refactor(layering): carve runtime-component out of runtime, break the Arc<Runtime> cycle (PR 6.1 + 5.3)
Carve the `component` module out of the `runtime` crate body into a sibling
`runtime-component` crate (Phase 6 PR 6.1, paired with the PR 5.3 config-spec
split). This breaks the `Arc<Runtime>` cycle so the config-only cores compile
in parallel with `runtime` instead of inside its single ~1127s codegen unit.

Split each of `Dataset`/`Catalog`/`View` into:
- a config-only `*Spec` (`DatasetSpec`/`CatalogSpec`/`ViewSpec`) in the new
  `runtime-component` crate, holding every declared field EXCEPT the runtime
  handles, plus all pure-config methods and the config submodules (`access`,
  `column`, `acceleration`, `declared_type`, `declared_schema`, `metadata`,
  `replication`, `schema_inference`); `runtime-component` names nothing from
  `runtime` (no `Arc<Runtime>`/`App`/`secrets`/`datafusion`), so it sits below
  `runtime` and the cycle is gone.
- an `Arc<Runtime>`-bound wrapper `{ spec, app, runtime }` kept in `runtime`,
  which `Deref`/`DerefMut`s to the spec so every `dataset.acceleration`,
  `catalog.provider`, `view.columns`, `dataset.refresh_sql()`, etc. keeps
  compiling unchanged. The wrapper keeps `AccelerationSource`, the builders,
  `is_accelerator_initialized`, `parse_table_reference`, and the error-returning
  constructors (they reference runtime-crate errors/constants).

`AcceleratedTable` is NOT pushed below runtime; its two `DataConnector` hooks
(`on_accelerator_setup`/`on_accelerated_table_registration`) stay runtime-side.

`runtime/src/component` becomes a thin shim re-exporting the moved items via
`pub use runtime_component::…` so all `crate::component::…` paths still resolve
during the migration (de-shimmed in the follow-up commit).

- add `runtime-component` (workspace member, `[lints] workspace=true`,
  copyright header, layers.toml `runtime` tier)
- layering guard green (131 crates, no upward edges)
