+++ Cargo.toml
@@ -32,6 +32,7 @@ members = [
     "crates/runtime-object-store",
     "crates/package",
     "crates/runtime",
+    "crates/runtime-component",
     "crates/runtime-api-types",
     "crates/runtime-checkpoint-api",
     "crates/runtime-checkpoint-duckdb",
