+++ crates/runtime-component/src/access.rs
