+++ crates/runtime/src/component/catalog.rs
@@ -18,10 +18,17 @@ use app::App;
 use globset::{Glob, GlobSet, GlobSetBuilder};
 use snafu::prelude::*;
 use spicepod::{component::catalog as spicepod_catalog, param::Params};
+use std::ops::{Deref, DerefMut};
 use std::{collections::HashMap, sync::Arc};
 
-use super::{find_first_delimiter, validate_identifier};
-use crate::{Runtime, component::access::AccessMode};
+use crate::Runtime;
+use crate::component::access::AccessMode;
+
+// Config-only spec + config types live in `runtime-component`; re-export for
+// path compatibility (`crate::component::catalog::CatalogSpec`, etc.).
+pub use runtime_component::catalog::{
+    CatalogAcceleration, CatalogAccelerationEngine, CatalogRefreshMode, CatalogSpec,
+};
 
 #[derive(Debug, Snafu)]
 pub enum Error {
@@ -36,69 +43,27 @@ pub enum Error {
 
 pub type Result<T> = std::result::Result<T, Error>;
 
-/// Acceleration configuration for an entire [`Catalog`]. See
-/// [`spicepod_catalog::CatalogAcceleration`] for the user-facing schema this
-/// mirrors.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct CatalogAcceleration {
-    pub engine: CatalogAccelerationEngine,
-    pub refresh_mode: CatalogRefreshMode,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub enum CatalogAccelerationEngine {
-    #[default]
-    Cayenne,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum CatalogRefreshMode {
-    Changes,
-}
-
-impl From<spicepod_catalog::CatalogAccelerationEngine> for CatalogAccelerationEngine {
-    fn from(engine: spicepod_catalog::CatalogAccelerationEngine) -> Self {
-        match engine {
-            spicepod_catalog::CatalogAccelerationEngine::Cayenne => {
-                CatalogAccelerationEngine::Cayenne
-            }
-        }
-    }
-}
-
-impl From<spicepod_catalog::CatalogRefreshMode> for CatalogRefreshMode {
-    fn from(refresh_mode: spicepod_catalog::CatalogRefreshMode) -> Self {
-        match refresh_mode {
-            spicepod_catalog::CatalogRefreshMode::Changes => CatalogRefreshMode::Changes,
-        }
-    }
-}
-
-impl From<spicepod_catalog::CatalogAcceleration> for CatalogAcceleration {
-    fn from(acceleration: spicepod_catalog::CatalogAcceleration) -> Self {
-        CatalogAcceleration {
-            engine: acceleration.engine.into(),
-            refresh_mode: acceleration.refresh_mode.into(),
-        }
-    }
-}
-
+/// `Arc<Runtime>`-bound wrapper over a [`CatalogSpec`]. Derefs to the spec so
+/// `catalog.provider`, `catalog.acceleration`, etc. keep working unchanged.
 #[derive(Clone)]
 pub struct Catalog {
-    pub provider: String,
-    pub catalog_id: Option<String>,
-    pub from: String,
-    pub name: String,
-    pub access: AccessMode,
-    pub(crate) orig_include: Vec<String>,
-    pub include: Option<GlobSet>,
-    pub(crate) orig_exclude: Vec<String>,
-    pub exclude: Option<GlobSet>,
-    pub params: HashMap<String, String>,
-    pub dataset_params: HashMap<String, String>,
-    pub acceleration: Option<CatalogAcceleration>,
+    pub spec: CatalogSpec,
     pub app: Arc<App>,
     pub runtime: Arc<Runtime>,
+}
+
+impl Deref for Catalog {
+    type Target = CatalogSpec;
+
+    fn deref(&self) -> &Self::Target {
+        &self.spec
+    }
+}
+
+impl DerefMut for Catalog {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.spec
+    }
 }
 
 impl std::fmt::Debug for Catalog {
@@ -123,14 +88,7 @@ impl std::fmt::Debug for Catalog {
 
 impl PartialEq for Catalog {
     fn eq(&self, other: &Self) -> bool {
-        self.from == other.from
-            && self.name == other.name
-            && self.access == other.access
-            && self.orig_include == other.orig_include
-            && self.orig_exclude == other.orig_exclude
-            && self.params == other.params
-            && self.dataset_params == other.dataset_params
-            && self.acceleration == other.acceleration
+        self.spec == other.spec
     }
 }
 
@@ -143,61 +101,6 @@ impl Catalog {
     #[must_use]
     pub fn runtime(&self) -> Arc<Runtime> {
         Arc::clone(&self.runtime)
-    }
-
-    /// Returns the catalog provider - the first part of the `from` field before the first '://', ':', or '/'.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use runtime::component::catalog::Catalog;
-    ///
-    /// let catalog = Catalog::new("foo:bar", "bar");
-    ///
-    /// assert_eq!(catalog.provider, "foo".to_string());
-    /// ```
-    ///
-    /// ```
-    /// use runtime::component::catalog::Catalog;
-    ///
-    /// let catalog = Catalog::new("foo", "bar");
-    ///
-    /// assert_eq!(catalog.provider, "foo".to_string());
-    /// ```
-    #[must_use]
-    fn provider(from: &str) -> &str {
-        match find_first_delimiter(from) {
-            Some((0, _)) | None => from,
-            Some((pos, _)) => &from[..pos],
-        }
-    }
-
-    /// Returns the catalog id - the second part of the `from` field after the first `:`.
-    /// This is optional and will return the default catalog from the provider if not set.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use runtime::component::catalog::Catalog;
-    ///
-    /// let catalog = Catalog::new("foo:bar", "bar");
-    ///
-    /// assert_eq!(catalog.catalog_id, Some("bar".to_string()));
-    /// ```
-    ///
-    /// ```
-    /// use runtime::component::catalog::Catalog;
-    ///
-    /// let catalog = Catalog::new("foo", "bar");
-    ///
-    /// assert_eq!(catalog.catalog_id, None);
-    /// ```
-    #[must_use]
-    fn catalog_id(from: &str) -> Option<&str> {
-        match find_first_delimiter(from) {
-            Some((pos, len)) => Some(&from[pos + len..]),
-            None => None,
-        }
     }
 }
 
@@ -241,13 +144,13 @@ impl TryFrom<spicepod_catalog::Catalog> for CatalogBuilder {
     type Error = crate::Error;
 
     fn try_from(catalog: spicepod_catalog::Catalog) -> std::result::Result<Self, Self::Error> {
-        let provider = Catalog::provider(&catalog.from);
-        let catalog_id = Catalog::catalog_id(&catalog.from).map(String::from);
+        let provider = CatalogSpec::provider(&catalog.from);
+        let catalog_id = CatalogSpec::catalog_id(&catalog.from).map(String::from);
 
         let include = compile_globset(&catalog.include)?;
         let exclude = compile_globset(&catalog.exclude)?;
 
-        validate_identifier(&catalog.name).context(crate::ComponentSnafu)?;
+        crate::component::validate_identifier(&catalog.name).context(crate::ComponentSnafu)?;
 
         if catalog
             .name
@@ -311,7 +214,7 @@ impl TryFrom<spicepod_catalog::Catalog> for CatalogBuilder {
 impl CatalogBuilder {
     #[expect(clippy::result_large_err)]
     pub fn try_new(from: String, name: &str) -> std::result::Result<Self, crate::Error> {
-        validate_identifier(name).context(crate::ComponentSnafu)?;
+        crate::component::validate_identifier(name).context(crate::ComponentSnafu)?;
 
         if name.eq_ignore_ascii_case(crate::datafusion::SPICE_DEFAULT_CATALOG) {
             return Err(crate::Error::ComponentError {
@@ -321,8 +224,8 @@ impl CatalogBuilder {
             });
         }
 
-        let provider = Catalog::provider(from.as_str());
-        let catalog_id = Catalog::catalog_id(from.as_str()).map(String::from);
+        let provider = CatalogSpec::provider(from.as_str());
+        let catalog_id = CatalogSpec::catalog_id(from.as_str()).map(String::from);
 
         Ok(CatalogBuilder {
             provider: provider.to_string(),
@@ -365,18 +268,20 @@ impl CatalogBuilder {
         })?;
 
         let catalog = Catalog {
-            provider: self.provider,
-            catalog_id: self.catalog_id,
-            from: self.from,
-            name: self.name,
-            access: self.access,
-            orig_include: self.orig_include,
-            include: self.include,
-            orig_exclude: self.orig_exclude,
-            exclude: self.exclude,
-            params: self.params,
-            dataset_params: self.dataset_params,
-            acceleration: self.acceleration,
+            spec: CatalogSpec {
+                provider: self.provider,
+                catalog_id: self.catalog_id,
+                from: self.from,
+                name: self.name,
+                access: self.access,
+                orig_include: self.orig_include,
+                include: self.include,
+                orig_exclude: self.orig_exclude,
+                exclude: self.exclude,
+                params: self.params,
+                dataset_params: self.dataset_params,
+                acceleration: self.acceleration,
+            },
             app,
             runtime,
         };
