+++ crates/runtime/src/component/dataset/builder.rs
@@ -272,30 +272,32 @@ impl DatasetBuilder {
             })?;
 
         let dataset = Dataset {
-            from: self.from,
-            name: self.name,
-            access: self.access,
-            params: self.params,
-            metadata: self.metadata,
-            columns: self.columns,
-            schema,
-            has_metadata_table: self.has_metadata_table,
-            replication: self.replication,
-            time_column: self.time_column,
-            time_format: self.time_format,
-            time_partition_column: self.time_partition_column,
-            time_partition_format: self.time_partition_format,
-            acceleration: self.acceleration,
-            embeddings: self.embeddings,
+            spec: super::DatasetSpec {
+                from: self.from,
+                name: self.name,
+                access: self.access,
+                params: self.params,
+                metadata: self.metadata,
+                columns: self.columns,
+                schema,
+                has_metadata_table: self.has_metadata_table,
+                replication: self.replication,
+                time_column: self.time_column,
+                time_format: self.time_format,
+                time_partition_column: self.time_partition_column,
+                time_partition_format: self.time_partition_format,
+                acceleration: self.acceleration,
+                embeddings: self.embeddings,
+                unsupported_type_action: self.unsupported_type_action,
+                on_schema_change: self.on_schema_change,
+                ready_state: self.ready_state,
+                metrics: self.metrics,
+                vectors: self.vectors,
+                full_text_search: self.full_text_search,
+                check_availability: self.check_availability,
+            },
             app,
-            unsupported_type_action: self.unsupported_type_action,
-            on_schema_change: self.on_schema_change,
-            ready_state: self.ready_state,
-            metrics: self.metrics,
             runtime,
-            vectors: self.vectors,
-            full_text_search: self.full_text_search,
-            check_availability: self.check_availability,
         };
 
         Ok(dataset)
